+++ HISTORY.rst
@@ -3,6 +3,15 @@
 Release History
 ---------------
 
+1.0.3 (2016-xx-xx)
+++++++++++++++++++
+* Added the ability to request the namespace for a TR64 action
+* Added interface ID's to all actions
+* Added a factory method to DeviceTR64 which takes an URL to the device definitions
+* Changed the declaration of getGenericAssociatedDeviceInfo for consistency reasons
+* Changed for all Wifi actions that the interface id is now optional
+* Fix an issue with relative URL's in the device definition
+
 1.0.1 (2016-01-04)
 ++++++++++++++++++